refactored SpeechInputSource
